requirements.txt
discord.py
python-dotenv
dateparser
parsedatetime
tzlocal